EventName saves - localdate/baby doesn't
i have to figure out how to do localdate.isEmpty() since its not a string. same with Baby. Its not a string.

Edit for Baby doesnt work so i been trying some code for that. I think im making progress with it because the servlet is sending it to the right place. it just doesn't seem to be getting the input. it gets a " " instead